Move Globals Mode template compiler initializer into ember-template-compiler
